Creating receipt works to PDF and EXCEL

+++ Phone Company sln/Common/Models/Package.cs
@@ -13,7 +13,7 @@ namespace Common.Models
         private double totalPrice;
         private int maxMinute;
         private double fixedPrice;
-        private double disscountPrecentage;
+        private int disscountPrecentage;
         private bool mostCallNumber;
         private bool insideFamilyCall;
         private int selectedNumberId;
